+++ peer/pendingheap/list_test.go
@@ -45,10 +45,6 @@ import (
 	"go.uber.org/zap/zaptest"
 )
 
-var (
-	_noContextDeadlineError = yarpcerrors.Newf(yarpcerrors.CodeInvalidArgument, `"fewest-pending-requests" peer list can't wait for peer without a context deadline`)
-)
-
 func newNotRunningError(err string) error {
 	return yarpcerrors.FailedPreconditionErrorf(`"fewest-pending-requests" peer list is not running: %s`, err)
 }
@@ -478,17 +474,6 @@ func TestPeerHeapList(t *testing.T) {
 					Wait: 20 * time.Millisecond,
 				},
 				ChooseAction{ExpectedPeer: "1"},
-			},
-			expectedRunning: true,
-		},
-		{
-			msg: "no blocking with no context deadline",
-			peerListActions: []PeerListAction{
-				StartAction{},
-				ChooseAction{
-					InputContext: context.Background(),
-					ExpectedErr:  _noContextDeadlineError,
-				},
 			},
 			expectedRunning: true,
 		},